Renamed file, whitespace fix and comment shift
